backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7
python-decouple==3.8
django-filter==23.5
drf-spectacular==0.26.5
